+++ src/main/java/com/nisovin/shopkeepers/config/migration/ConfigMigration.java
@@ -4,5 +4,14 @@ import org.bukkit.configuration.Configuration;
 
 public interface ConfigMigration {
 
+	/**
+	 * Applies the config migration.
+	 * <p>
+	 * If an issue prevents the migration of individual config entries, it is usually preferred to log a warning and
+	 * continue the migration, instead of aborting the migration or throwing an exception.
+	 * 
+	 * @param config
+	 *            the current config
+	 */
 	public void apply(Configuration config);
 }